delete REINFORCE code (#68)
