+++ src/mcts/__init__.py
@@ -1,4 +1,4 @@
-from .model import MCTSModel, PolicyMCTSModel, PolicyMCTSModelConfig, PPOMCTSModel, PPOMCTSModelConfig, ReinforceMCTSModel, ReinforceMCTSModelConfig
+from .model import MCTSModel, PolicyMCTSModel, PolicyMCTSModelConfig, PPOMCTSModel, PPOMCTSModelConfig
 from .config import FixedMCTSConfig, MCTSConfig, NNMCTSConfig, WideningMCTSConfig
 from .base import MCTS, Node, Transition
 from .nn import NNMCTS, NNMCTSMode
