+++ src/mcts/model/__init__.py
@@ -1,5 +1,4 @@
-from .config import PolicyMCTSModelConfig, PPOMCTSModelConfig, ReinforceMCTSModelConfig
-from .reinforce import ReinforceMCTSModel
+from .config import PolicyMCTSModelConfig, PPOMCTSModelConfig
 from .policy import PolicyMCTSModel
 from .ppo import PPOMCTSModel
 from .base import MCTSModel
